requirements.txt
requests
feedparser
newspaper3k
google-api-python-client
google-auth-oauthlib
